Chore: moved UserDto package
